Incorporating the callhome dataset.
In order to understand how the code work, we are currently working with the callhome dataset that can be found in https://huggingface.co/datasets/talkbank/callhome I don't know this is the dataset of the original EEND paper, or some kind of remaster hahaha.

We are going to see if the code is working and we can test the repo with this dataset.