Callhome-test problem solved.
I solved some of the initial problems realted with the dataset. This CALLHOME dataset is a little difficult to make use, but it seems we are starting to see the light.

For accessing an audio file, we need to go to the dataset, then data key, then audio key and then there are multiple samples.

As we have seen in the dataset we have:

As we can see, here we are not using the dataloader. We may need to explore this possibility further.

## Next steps:

Now we have an error because we don't have a dev_set. We need to find a way to define this dev_set by either doing a partition of the total data or finding any other solution. The problem I fear is that we may not have enough data for doing train, validation and set.